Move custom filters to their own module
