+++ app/server/static/js/document_classification.js
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import annotationMixin from './mixin';
 import HTTP from './http';
-import simpleShortcut from './filter';
+import { simpleShortcut } from './filter';
 
 Vue.use(require('vue-shortkey'), {
   prevent: ['input', 'textarea'],
